Added Presets for Classes, and added Dummy Sizes to the Dummy Class
